requirements.txt
discord.py==2.2.2
python-dotenv
py-cord===2.2.2
